Move CgcloudTestCase to lib.test so Toil can use it
