+++ core/src/cgcloud/core/__init__.py
@@ -48,4 +48,3 @@ def command_classes( ):
     return __fail_deprecated( sorted( locals( ).values( ), key=lambda cls: cls.__name__ ) )
 
 
-test_namespace_suffix_length = 11
