+++ lib/src/cgcloud/lib/__init__.py
@@ -1,3 +1,5 @@
 from bd2k.util.d64 import D64
 
 aws_d64 = D64( '.-' )  # hopefully the dot is supported for all AWS resource names
+
+test_namespace_suffix_length = 11
